refactor(cli): index.rs を mod.rs + runtime.rs + audit.rs に分割 (SPEC-1942 SC-027)
SPEC-1942 SC-027 (200-600 行) を index family で完全達成。 cli/index.rs
(751 行) を 3 sub-file に分割:

- cli/index/mod.rs (360 行): parse + run dispatch + run_status + run_rebuild +
  runtime_error + tests。 SC-027 範囲内。
- cli/index/runtime.rs (227 行): IndexContext + resolve_index_context +
  project_index_python_path + run_runner_status + RebuildAction +
  rebuild_actions + run_runner_rebuild + parse_runner_json +
  render_index_status + format_runner_failure + io_error。 Python runner
  integration を集約。
- cli/index/audit.rs (218 行): audit_log_dir + audit_status + AuditHealth +
  audit_health + audit_status_failure + audit_rebuild_start +
  audit_rebuild_result + audit_runner_progress + base_audit_fields +
  append_index_audit_event。 JSONL audit logging を集約。

mod.rs は use audit::{...} と use runtime::{...} で必要 helper のみ取り込み。
audit.rs は use super::runtime::{format_runner_failure, IndexContext} で
runtime に依存、 全体的に依存方向が一方向になる。

機能・契約・test 結果は不変 (lib 345 全 pass、 clippy 0 / fmt clean)。
